Refactor: extract onboard Group D (callable registry + upload) into base
Move the per-callable_id registration surface + chip-callable upload from
a2a3 and a5 onboard DeviceRunners into the shared DeviceRunnerBase
(-346 net lines):

- Methods: upload_chip_callable_buffer, prepare_orch_so, register_callable,
  register_callable_host_orch, unregister_callable, has_callable,
  bind_callable_to_runtime, aicpu_dlopen_count(), host_dlopen_count().
- State: callables_, orch_so_dedup_, aicpu_seen_callable_ids_,
  aicpu_dlopen_total_, host_dlopen_total_, chip_callable_buffers_.
- Inner struct types: CallableState, OrchSoBuffer, ChipCallableBuffer.

upload_chip_callable_buffer now uses the shared
compute_chip_callable_layout + patch_chip_callable_scratch_for_device
helpers from src/common/task_interface/chip_callable_layout.h (a2a3
already used them; a5's inline equivalent is replaced).

Verbatim move otherwise — no behavioral change. nm -D confirms the
methods now resolve to DeviceRunnerBase symbols on both arches' .so.
Local a2a3 onboard smoke (dummy_task, alternating_matmul_add,
prepared_callable suite) — 8/8 passed in 15s.
